+++ nextgisweb/file_upload/api.py
@@ -177,8 +177,11 @@ def _collection_post_tus(request, *, comp: FileUploadComponent):
         size=upload_length,
         name=upload_metadata.get("name"),
         mime_type=upload_metadata.get("mime_type"),
-        incomplete=True,
+        incomplete=upload_length > 0,
     )
+
+    if upload_length == 0 and fupload.mime_type is None:
+        fupload.mime_type = "application/octet-stream"
 
     fupload.data_path.touch()
     fupload.write_meta()
@@ -333,7 +336,16 @@ def _tus_decode_upload_metadata(value):
 
 
 def setup_pyramid(comp, config):
-    tus_request_headers = ("Upload-Offset", "Upload-Length", "Tus-Resumable")
+    tus_cors_headers = dict(
+        request=(
+            "Upload-Offset",
+            "Upload-Length",
+            "Upload-Defer-Length",
+            "Upload-Metadata",
+            "Tus-Resumable",
+        ),
+        response=("Location",),
+    )
 
     # TODO: Remove legacy route: Formbuilder
     config.add_route(
@@ -350,7 +362,7 @@ def setup_pyramid(comp, config):
         options=collection_options,
         post=collection_post,
         put=collection_put,
-        cors_headers=tus_request_headers,
+        cors_headers=tus_cors_headers,
     )
 
     config.add_route(
@@ -361,5 +373,5 @@ def setup_pyramid(comp, config):
         get=item_get,
         patch=item_patch_tus,
         delete=item_delete,
-        cors_headers=tus_request_headers,
+        cors_headers=tus_cors_headers,
     )