+++ nextgisweb/pyramid/nodepkg/i18n/testentry.ts
@@ -3,7 +3,6 @@ import { assert } from "chai";
 
 import {
     gettextf,
-    ngettext,
     ngettextf,
     npgettext,
     npgettextf,
@@ -60,12 +59,71 @@ describe("Gettext implementation", () => {
         const f = (num: number) => npgettext("test", "unit", "units", num);
         assert.notStrictEqual(f(1), f(2));
     });
-
-    it("npgettextf implementation", () => {
-        const num = 3;
-        assert.equal(
-            npgettextf("test", "{} unit", "{} units", num)(num + ` edited`),
-            "3 edited units"
-        );
-    });
 });
+
+const describeInterpolation = (
+    family: string,
+    sFn: typeof gettextf,
+    nFn: typeof ngettextf
+) => {
+    const pluralUnit = (n: number) => npgettext("test", "unit", "units", n);
+
+    describe(`String iterpolation with ${family}`, () => {
+        it("handles positional arguments", () => {
+            const f = sFn("Hello, {1} {0}");
+            assert.equal(f("Dent", "Arthur"), "Hello, Arthur Dent");
+        });
+
+        it("handles named arguments", () => {
+            const f = sFn("Hello, {first} {last}");
+            assert.equal(
+                f({ first: "Arthur", last: "Dent" }),
+                "Hello, Arthur Dent"
+            );
+        });
+
+        it("unescapes trivial curly braces", () => {
+            const f = sFn("{{Hello}}, {0}");
+            assert.equal(f("Arthur"), "{Hello}, Arthur");
+        });
+
+        it("unescapes nested curly braces", () => {
+            const f = sFn("Hello, {{{0}}}");
+            assert.equal(f("Arthur"), "Hello, {Arthur}");
+        });
+
+        it("skips unused arguments", () => {
+            const f = sFn("Hello, Arthur");
+            assert.equal(f("Dent"), "Hello, Arthur");
+        });
+
+        it("throws error on missing arguments", () => {
+            const f = sFn("Hello, {1} {0}");
+            assert.throws(() => f("Dent"));
+        });
+
+        it("does simple pluralization", () => {
+            const f = (n: number) => {
+                return nFn("One deleted", "{} deleted", n);
+            };
+            assert.equal(f(1)(1), "One deleted");
+            assert.equal(f(2)(2), "2 deleted");
+        });
+
+        it("does complex pluralization", () => {
+            const f = (n: number) => {
+                return nFn("One {1} deleted", "{0} {1} deleted", n);
+            };
+            assert.equal(f(1)(1, pluralUnit(1)), "One unit deleted");
+            assert.equal(f(2)(2, pluralUnit(2)), "2 units deleted");
+        });
+    });
+};
+
+describeInterpolation("gettext", gettextf, ngettextf);
+
+describeInterpolation(
+    "pgettext",
+    (...args) => pgettextf("test", ...args),
+    (...args) => npgettextf("test", ...args)
+);