+++ nextgisweb/wmsclient/model.py
@@ -23,9 +23,9 @@ from nextgisweb.resource import (
     ConnectionScope,
     CRUTypes,
     DataScope,
-    DataStructureScope,
     Resource,
     ResourceGroup,
+    ResourceScope,
     SAttribute,
     SColumn,
     Serializer,
@@ -224,7 +224,7 @@ class Layer(Base, Resource, SpatialLayerMixin):
     identity = "wmsclient_layer"
     cls_display_name = gettext("WMS layer")
 
-    __scope__ = (DataStructureScope, DataScope)
+    __scope__ = DataScope
 
     connection_id = db.Column(db.ForeignKey(Resource.id), nullable=False)
     wmslayers = db.Column(db.Unicode, nullable=False)
@@ -313,8 +313,8 @@ class LayerSerializer(Serializer, apitype=True):
     identity = Layer.identity
     resclass = Layer
 
-    connection = SResource(read=DataStructureScope.read, write=DataStructureScope.write)
-    wmslayers = SColumn(read=DataStructureScope.read, write=DataStructureScope.write)
-    imgformat = SColumn(read=DataStructureScope.read, write=DataStructureScope.write)
-    vendor_params = VendorParamsAttr(read=DataStructureScope.read, write=DataStructureScope.write)
-    srs = SRelationship(read=DataStructureScope.read, write=DataStructureScope.write)
+    connection = SResource(read=ResourceScope.read, write=ResourceScope.update)
+    wmslayers = SColumn(read=ResourceScope.read, write=ResourceScope.update)
+    imgformat = SColumn(read=ResourceScope.read, write=ResourceScope.update)
+    vendor_params = VendorParamsAttr(read=ResourceScope.read, write=ResourceScope.update)
+    srs = SRelationship(read=ResourceScope.read, write=ResourceScope.update)