+++ CHANGES.rst
@@ -1,6 +1,13 @@
 Changes
 =======
 
+4.9.0
+-----
+
+- Fix uploading of zero-length files.
+- TUS uploader is available using CORS.
+- Data structure permission scope removed.
+
 4.8.0
 -----